requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-telegram-bot==20.7
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
passlib[bcrypt]==1.7.4
mutagen==1.47.0
pillow==10.1.0
pydantic-settings==2.0.3 